requirements.txt
snowflake-connector-python==2.6.2
snowflake-sqlalchemy==1.3.2
ruamel.yaml
scikit-learn
nbdev
fastcore
xgboost
hyperopt==0.2.7
pyarrow==5.0.0
fastai
scikit-plot
rfpimp
flake8
azure-storage-file-datalake==12.4.0
azure-core==1.13.0
azure-storage-blob==12.8.1
pysftp
jupyter-client==6.1.12